Rename git package
