+++ clowder/data/model/project.py
@@ -10,7 +10,7 @@ from typing import List, Optional, Union
 
 import clowder.util.formatting as fmt
 from clowder.error import ClowderError, ClowderErrorType
-from clowder.git.util import (
+from clowder.git_project.util import (
     format_git_branch,
     format_git_tag
 )
