+++ clowder/git_project/project_repo_impl.py
