+++ clowder/git_project/util.py
